Reduce MSRV and add a MSRV build job
The namespaced features with the `dep` prefix was stabilized with 1.60,
and the MSRV can be reduced to 1.27.2 by avoiding it.

It's not necessary to define a feature to enable a single optional
dependencies.

The build job makes it possible to keep the MSRV in check and bump the
version deliberately when needed.